2023.05.16 merged

+++ src/main/java/com/cook/mymealkit/mapper/BuyMapper.java
@@ -14,13 +14,13 @@ public interface BuyMapper {
 	public void insertUserBuy(BuyUserVO vo); 				// 회원구매 등록
 	public List<BuyUserVO> userBuyList();  					// 회원구매 목록
 	public List<BuyUserVO> bListByUserId(String user_id); 	// 회원구매 조회
-	public Long getMaxBno();								// bno 최대값 조회
+	public int getMaxBno();								// bno 최대값 조회
 	
 	
 	/* 비회원 */
 	public void insertGuestBuy(BuyGuestVO vo); 				// 게스트구매 등록
 	public List<BuyGuestVO> guestBuyList();					// 게스트구매 목록
 	public List<BuyGuestVO> bListByGuestId(String buy_no);	// 게스트구매 조회
-	public Long getMaxGno();								// gno 최대값 조회
+	public int getMaxGno();								// gno 최대값 조회
 	
 }
